Preserve accu in %rbx when necessary.
Since %rbx is free now, we can use it to preserve the accu when necessary
(in particular within the compile trampoline). This also simplifies the
compile trampoline.

A bug in caml_jit_rt_makeshrblock was also fixed, which preserved %rax
in an invalid stack location. Now also uses %rbx for this purpose.